requirements.txt
Flask
Flask-RESTful
prometheus-client
Flask-SQLAlchemy
psycopg2-binary
